reminder_service/requirements.txt
Flask==2.1.1
functions-framework==3.0.0
line-bot-sdk==1.20.0
Werkzeug==2.0.3
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-scheduler
pytz
requests